Move methods to MasterJobContext
